src/requirements.txt
kurigram
pytgcrypto
aiosqlite
tortoise-orm
pydantic-settings